requirements.txt
# Heart Attack Anomaly Detection with XAI
# Production-ready dependencies

# Core Scientific Computing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Explainability
shap>=0.42.0

# API and Web
requests>=2.31.0
huggingface-hub>=0.20.0

# Utilities
python-dotenv>=1.0.0

# Optional: For visualization (not required for production)
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: For Jupyter notebooks (if needed for exploration)
# jupyter>=1.0.0
# ipykernel>=6.25.0
